+++ library/sbi-rt/CHANGELOG.md
@@ -20,6 +20,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - rt: add SSE extension support to SBI implementation.
 - feat(rt): add MPXY extension support to SBI runtime library.
 - lib: add `raw` module for raw SBI calls; they may be used in runtime libraries for custom SBI extensions.
+- lib: add raw RISC-V SBI calls with 3 to 6 parameters.
 
 ### Modified